pyproject.toml
[tool.poetry]
name = "scbl-utils"
version = "2.0.0"
description = "A set of command-line utilities that facilitate data processing at the Single Cell Biology Lab at the Jackson Laboratory."
authors = ["Ahmed Said <ahmed.said@jax.org>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/TheJacksonLaboratory/scbl-utils/"
documentation = "https://github.com/TheJacksonLaboratory/scbl-utils/"

[tool.poetry.scripts]
scbl-utils = 'scbl_utils.main:main'

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
PyYAML = "^6.0.1"
jsonschema = "^4.21.1"
email-validator = "^2.1.0.post1"
sqlalchemy = "^2.0.25"
gspread = "^5.12.4"
pandas = "^2.2.0"
pydantic = "^2.6.1"
fire = "^0.5.0"
rich = "^13.7.0"
pyarrow = "^15.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
skip-string-normalization = true